move common methods to wfs client
